Shift audit log functionality to logging service
