+++ src/main/java/teammates/logic/api/LogsProcessor.java
@@ -1,8 +1,11 @@
 package teammates.logic.api;
 
+import java.time.Instant;
 import java.util.List;
 
 import teammates.common.datatransfer.ErrorLogEntry;
+import teammates.common.datatransfer.FeedbackSessionLogEntry;
+import teammates.common.exception.LogServiceException;
 import teammates.common.util.Config;
 import teammates.logic.core.GoogleCloudLoggingService;
 import teammates.logic.core.LocalLoggingService;
@@ -31,6 +34,23 @@ public class LogsProcessor {
      */
     public List<ErrorLogEntry> getRecentErrorLogs() {
         return service.getRecentErrorLogs();
+    }
+
+    /**
+     * Creates a feedback session log.
+     */
+    public void createFeedbackSessionLog(String courseId, String email, String fsName, String fslType)
+            throws LogServiceException {
+        service.createFeedbackSessionLog(courseId, email, fsName, fslType);
+    }
+
+    /**
+     * Gets the feedback session logs as filtered by the given parameters.
+     * @param email Can be null
+     */
+    public List<FeedbackSessionLogEntry> getFeedbackSessionLogs(String courseId, String email,
+            Instant startTime, Instant endTime) throws LogServiceException {
+        return service.getFeedbackSessionLogs(courseId, email, startTime, endTime);
     }
 
 }
