+++ src/main/java/teammates/logic/core/LocalLoggingService.java
@@ -1,9 +1,12 @@
 package teammates.logic.core;
 
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 
 import teammates.common.datatransfer.ErrorLogEntry;
+import teammates.common.datatransfer.FeedbackSessionLogEntry;
+import teammates.common.exception.LogServiceException;
 
 /**
  * Holds functions for operations related to logs reading/writing in local dev environment.
@@ -17,6 +20,19 @@ public class LocalLoggingService implements LogService {
 
     @Override
     public List<ErrorLogEntry> getRecentErrorLogs() {
+        // Not supported in dev server
+        return new ArrayList<>();
+    }
+
+    @Override
+    public void createFeedbackSessionLog(String courseId, String email, String fsName, String fslType)
+            throws LogServiceException {
+        // Not supported in dev server
+    }
+
+    @Override
+    public List<FeedbackSessionLogEntry> getFeedbackSessionLogs(String courseId, String email,
+            Instant startTime, Instant endTime) throws LogServiceException {
         // Not supported in dev server
         return new ArrayList<>();
     }
