+++ src/main/java/teammates/ui/webapi/ActionFactory.java
@@ -119,7 +119,7 @@ public class ActionFactory {
         map(ResourceURIs.INSTRUCTOR, POST, CreateInstructorAction.class);
 
         // Logging and tracking
-        map(ResourceURIs.TRACK_SESSION, GET, CreateFeedbackSessionLogAction.class);
+        map(ResourceURIs.TRACK_SESSION, POST, CreateFeedbackSessionLogAction.class);
         map(ResourceURIs.SESSION_LOGS, GET, GetFeedbackSessionLogsAction.class);
 
         // Cron jobs; use GET request
