--- src/test/java/teammates/logic/core/LogsLogicTest.java
@@ -1,25 +1,0 @@
-package teammates.logic.core;
-
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-/**
- * SUT: {@link LogsLogic}.
- */
-public class LogsLogicTest extends BaseLogicTest {
-
-    @Override
-    protected void prepareTestData() {
-        // test data is refreshed before each test case
-    }
-
-    @BeforeMethod
-    public void refreshTestData() {
-        // TODO: refresh test data
-    }
-
-    @Test
-    public void testAll() throws Exception {
-        // TODO: test all
-    }
-}
